+++ src/app/common/user/user-state.model.ts
@@ -1,0 +1,5 @@
+import {UserBusinessModel} from "../business-models/user.business-model";
+
+export interface UserStateModel {
+  user: UserBusinessModel | null;
+}